backend/pyproject.toml
[project]
name = "langear-backend"
version = "0.1.0"
description = "LanGear MVP Backend - English Speaking Training Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "fsrs>=6.3.0",
    "google-generativeai>=0.8.0",
    "oss2>=2.18.0",
    "python-multipart>=0.0.6",
    "dashscope>=1.25.6",
    "aliyun-python-sdk-core>=2.16.0",
    "aliyun-python-sdk-sts>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]
